Created a start with the finish tile
I created a new Tile in the array and placed it on the map.txt (it has the value of 3 if you want to know where i placed it temporarily). When the startscreen is ready we can create some code in the if-statement that is checking if the player touches the finish tile and we can return the start screen and stop the game. Also the code that is checking for the finish tile (in the Map.Java file) is almost the same code for the collisions. So in the future we might check if we can do something about the code redundancy.